Last changes

+++ Infrastructure/TravelaFinalApp.Persistence/Implementations/DestinationService.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using TravelaFinalApp.Application.Dtos.DestinationDtos;
+using TravelaFinalApp.Application.Exceptions;
 using TravelaFinalApp.Application.Helpers;
 using TravelaFinalApp.Application.Interfaces;
 using TravelaFinalApp.Domain.Entities;
@@ -11,6 +12,8 @@ namespace TravelaFinalApp.Persistence.Implementations
     {
         public async Task CreateAsync(DestinationCreateDto destinationCreateDto)
         {
+            if (await destinationRepository.IsExist(d => d.DestinationPlace.ToLower().Trim() == destinationCreateDto.DestinationPlace.ToLower().Trim()))
+                throw new CustomException(400, "DestinationPlace", $" '{destinationCreateDto.DestinationPlace}' - with same name destination already exist!!");
             var destination= _mapper.Map<Destination>(destinationCreateDto);
             await destinationRepository.CreateAsync(destination);
             await destinationRepository.SaveChangesAsync();
@@ -20,26 +23,31 @@ namespace TravelaFinalApp.Persistence.Implementations
         {
             var existDestination=await destinationRepository.GetByIdAsync(id);
             if (existDestination == null)
-                throw new NullReferenceException("Destination is not found..");
+                throw new CustomException(404, "Id", "Data not found..");
             existDestination.IsDeleted=true;
             await destinationRepository.SaveChangesAsync();
         }
 
         public async Task<List<DestinationReturnDto>> GetAllAsync()
         {
-            return _mapper.Map<List<DestinationReturnDto>>(await destinationRepository.GetAllAsync());
+            return _mapper.Map<List<DestinationReturnDto>>(await destinationRepository.GetAllAsync(d=>!d.IsDeleted));
         }
 
         public async Task<DestinationReturnDto> GetByIdAsync(int id)
         {
-            return _mapper.Map<DestinationReturnDto>(await destinationRepository.GetByIdAsync(id));
+            var existDestination = await destinationRepository.GetByIdAsync(id);
+            if (existDestination == null)
+                throw new CustomException(404, "Id", "Data not found..");
+            return _mapper.Map<DestinationReturnDto>(existDestination);
         }
 
         public async Task UpdateAsync(int id, DestinationUpdateDto destinationUpdateDto)
         {
             var existDestination = await destinationRepository.GetByIdAsync(id);
             if (existDestination == null)
-                throw new NullReferenceException("Destination is not found..");
+                throw new CustomException(404, "Id", "Data not found..");
+            if (await destinationRepository.IsExist(d => d.DestinationPlace.ToLower().Trim() == destinationUpdateDto.DestinationPlace.ToLower().Trim()))
+                throw new CustomException(400, "DestinationPlace", $" '{destinationUpdateDto.DestinationPlace}' - with same name destination already exist!!");
             string path = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "images", existDestination.MainImage);
             FileHelper.DeleteFileFromRoute(path);
             _mapper.Map(destinationUpdateDto, existDestination);
