+++ Presentation/TravelaFinalApp.Presentation/Controllers/Admin/BlogController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using TravelaFinalApp.Application.Dtos.BlogDtos;
 using TravelaFinalApp.Application.Interfaces;
@@ -8,6 +9,7 @@ namespace TravelaFinalApp.Presentation.Controllers.Admin
 {
     [Route("api/admin/[controller]/[action]")]
     [ApiController]
+    [Authorize(Roles = "Admin")]
     public class BlogController(IBlogService blogService) : ControllerBase
     {
 
@@ -21,6 +23,8 @@ namespace TravelaFinalApp.Presentation.Controllers.Admin
         [HttpPut("{id}")]
         public async Task<IActionResult> Update([FromRoute]int id,BlogUpdateDto dto)
         {
+            if (id <= 0)
+                return BadRequest("Id can't be zero or negative");
             await blogService.UpdateAsync(id,dto);
             return Ok(new { Response = "Data successfully updated" });
         }
@@ -28,6 +32,8 @@ namespace TravelaFinalApp.Presentation.Controllers.Admin
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete([FromRoute]int id)
         {
+            if (id <= 0)
+                return BadRequest("Id can't be zero or negative");
             await blogService.DeleteAsync(id);
             return Ok(new { Response = "Data successfully deleted" });
         }
@@ -35,12 +41,14 @@ namespace TravelaFinalApp.Presentation.Controllers.Admin
         [HttpGet("")]
         public async Task<IActionResult> GetAll(int page, string? search)
         {
-            return Ok(await blogService.GetAllAsync());
+            return Ok(await blogService.GetAllAsync(page,search));
         }
 
         [HttpGet("{id}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (id <= 0)
+                return BadRequest("Id can't be zero or negative");
             return Ok(await blogService.GetByIdAsync(id));
         }
     }
