+++ Presentation/TravelaFinalApp.Presentation/Controllers/Admin/ServiceController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using TravelaFinalApp.Application.Dtos.ServiceDtos;
 using TravelaFinalApp.Application.Interfaces;
 
@@ -6,6 +7,7 @@ namespace TravelaFinalApp.Presentation.Controllers.Admin
 {
     [Route("api/admin/[controller]/[action]")]
     [ApiController]
+    [Authorize(Roles = "Admin")]
     public class ServiceController(IServiceService serviceService) : ControllerBase
     {
         [HttpPost("")]
@@ -18,6 +20,8 @@ namespace TravelaFinalApp.Presentation.Controllers.Admin
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+            if (id <= 0)
+                return BadRequest("Id can't be zero or negative");
             await serviceService.DeleteAsync(id);
             return Ok(new { Response = "Data successfully deleted.." });
         }
@@ -25,6 +29,8 @@ namespace TravelaFinalApp.Presentation.Controllers.Admin
         [HttpPut("{id}")]
         public async Task<IActionResult> Update([FromRoute]int id, [FromForm] ServiceUpdateDto serviceUpdateDto)
         {
+            if (id <= 0)
+                return BadRequest("Id can't be zero or negative");
             await serviceService.UpdateAsync(id, serviceUpdateDto);
             return Ok(new { Response = "Data successfully updated.." });
 
@@ -39,6 +45,8 @@ namespace TravelaFinalApp.Presentation.Controllers.Admin
         [HttpGet("{id}")]
         public async Task<IActionResult> GetById(int id)
         {
+            if (id <= 0)
+                return BadRequest("Id can't be zero or negative");
             return Ok(await serviceService.GetByIdAsync(id));
         }
     }
