+++ Presentation/TravelaFinalApp.Presentation/Controllers/Admin/SettingController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using TravelaFinalApp.Application.Dtos.SettingDtos;
 using TravelaFinalApp.Application.Interfaces;
 
@@ -6,6 +7,7 @@ namespace TravelaFinalApp.Presentation.Controllers.Admin
 {
     [Route("api/admin/[controller]/[action]")]
     [ApiController]
+    [Authorize(Roles = "Admin")]
     public class SettingController(ISettingService settingService) : ControllerBase
     {
         [HttpPost("")]
@@ -18,6 +20,8 @@ namespace TravelaFinalApp.Presentation.Controllers.Admin
         [HttpPut("{id}")]
         public async Task<IActionResult> Update([FromRoute]int id, SettingUpdateDto settingUpdateDto)
         {
+            if (id <= 0)
+                return BadRequest("Id can't be zero or negative");
             await settingService.UpdateAsync(id, settingUpdateDto);
             return Ok(new { Response = "Data updated successfully.." });
         }
@@ -25,6 +29,8 @@ namespace TravelaFinalApp.Presentation.Controllers.Admin
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete([FromRoute]int id)
         {
+            if (id <= 0)
+                return BadRequest("Id can't be zero or negative");
             await settingService.DeleteAsync(id);
             return Ok(new { Response = "Data deleted successfully" });
         }
